services/processor/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
exceptiongroup==1.3.1
fastapi==0.128.8
flatbuffers==25.12.19
h11==0.16.0
httptools==0.7.1
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.9.0
protobuf==6.33.5
pydantic==2.12.5
pydantic_core==2.41.5
PyMatting==1.1.15
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.36.2
rembg==2.0.61
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.24.0
scipy==1.13.1
starlette==0.49.3
sympy==1.14.0
tifffile==2024.8.30
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
